catalog/swatch-cropper/pyproject.toml
[project]
name = "swatch-cropper"
version = "0.1.0"
description = "Interactive multi-crop tool for fabric swatch cards"
requires-python = ">=3.9"
dependencies = [
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "pillow-heif>=0.16.0",
]

[project.scripts]
swatch-cropper = "swatch_cropper.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/swatch_cropper"]
